feat: move scope and provide usecase
